pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-standards-server"
version = "0.1.0"
description = "A Model Context Protocol server for intelligent access to development standards"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "William Zujkowski"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core dependencies
    "mcp>=0.1.0",  # Model Context Protocol SDK - critical dependency
    "pydantic>=2.0",
    "redis>=4.0.0",
    "PyYAML>=6.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.4.0",
    "tree-sitter>=0.20.0",
    "aiofiles>=23.0.0",
    "jsonschema>=4.0.0",
    "aiohttp>=3.9.0",
    "scikit-learn>=1.0.0",
    "numpy>=1.21.0",
    "nltk>=3.8.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.12.0",  # For better fuzzy matching performance
    "PyJWT>=2.8.0",  # For JWT authentication
    # Database dependencies with Python 3.12 compatibility
    "aioredis>=2.0.1,<3.0.0",  # Use compatible version
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "sqlalchemy>=2.0.0",
    # GitHub integration dependencies
    "PyGithub>=1.58.0",
    "requests>=2.25.0",
    # Performance monitoring and profiling dependencies
    "prometheus_client>=0.16.0",  # For metrics collection
    "pympler>=1.0.0",  # For memory profiling
    "memory-profiler>=0.60.0",  # For memory analysis
    "faiss-cpu>=1.7.0",  # For vector operations
    "psutil>=5.9.0",  # For system monitoring
    "msgpack>=1.0.0",  # For efficient serialization
]

[project.urls]
Homepage = "https://github.com/williamzujkowski/mcp-standards-server"
Repository = "https://github.com/williamzujkowski/mcp-standards-server"
Issues = "https://github.com/williamzujkowski/mcp-standards-server/issues"

[project.scripts]
mcp-standards = "cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "pytest-timeout>=2.1.0",
    "pytest-cov>=4.0.0",
    "coverage[toml]>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pip-audit>=2.6.0",
    "safety>=2.3.0",
]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "pytest-timeout>=2.1.0",
    "pytest-cov>=4.0.0",
    "coverage[toml]>=7.0",
    "memory-profiler>=0.60.0",
    "psutil>=5.9.0",
]
performance = [
    "memory-profiler>=0.60.0",
    "psutil>=5.9.0",
    "pympler>=1.0.0",
    "tracemalloc-ng>=1.0.0",
    "pytest-memprof>=0.2.0",
]
visualization = [
    "matplotlib>=3.5.0",  # For benchmark plots
]
full = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "faiss-cpu>=1.7.0",
    "annoy>=1.17.0",
    "matplotlib>=3.5.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
timeout = 300

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
concurrency = ["multiprocessing", "thread"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstract",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "B904", # raise from
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.md"]